qgswfsgetfeature: Properly check axis inversion in startGetFeature
The axis inversion needs to be checked on the output crs, not the
input one.

The existing test does not detect this issue because the requested
output crs is the same as the input one.
To cover this case, a new test with a output crs different from the
input one is added.