qgsogcutils: Fix axis inversion in geometryFromGML
`QgsOgcUtils::geometryFromGML` converts a GML to a QgsGeometry. in the
GML3 case, this checks if the X/Y coordinates need to be inverted only
if a context is provided and the destination crs is different from the
source crs.

However, the check for axis inversion should be done whether a context
is provided or not. This issue is fixed by the moving the `srsName`
and `swapXy` logic before the context check.